feat: Add hide transaction functionality
Introduces the ability to mark transactions as hidden. Hidden transactions will not appear in main lists (like Resumen or Gastos) but will still be included in calculations if relevant. This is useful for non-financial transactions or entries that should be tracked but not clutter the primary financial overview.

Also includes preparation for new 'Add' and 'Transferencia' pages by updating the `Page` type and related routing.